refactor: error handlers
